+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Fix config loading object-like checks.
+- Allow commander adapter.
+- Fix number ranges limited my safety.
+
 ## [3.2.0] - 2024-12-31
 
 - Add `Config.setData`.